fix(pr): address post-a60d898 review — PerspectiveBadge, score utils, description model env, shared Section
